+++ src/global/progresscounter.h
@@ -3,6 +3,8 @@
 // Copyright (C) 2019 The MMapper Authors
 // Author: Thomas Equeter <waba@waba.be> (Waba)
 
+#include <cstddef>
+
 #include <QObject>
 #include <QtGlobal>
 
@@ -11,7 +13,7 @@ class ProgressCounter final : public QObject
     Q_OBJECT
 
 private:
-    uint32_t m_totalSteps = 0u, m_steps = 0u, m_percentage = 0u;
+    size_t m_totalSteps = 0u, m_steps = 0u, m_percentage = 0u;
 
 public:
     ProgressCounter() = default;
@@ -19,10 +21,10 @@ public:
     ~ProgressCounter() final;
 
 public:
-    void step(uint32_t steps = 1u);
-    void increaseTotalStepsBy(uint32_t steps);
+    void step(size_t steps = 1u);
+    void increaseTotalStepsBy(size_t steps);
     void reset();
 
 signals:
-    void sig_onPercentageChanged(uint32_t);
+    void sig_onPercentageChanged(size_t);
 };
